Fixes/improvements to Steam(works) interactions/subscription handler
