+++ util/generic.py
@@ -8,15 +8,15 @@ from requests import post as requests_post
 from model.dialogue import show_information, show_warning
 
 
-def chunks(list: list, limit: int):
+def chunks(_list: list, limit: int):
     """
     Split list into chunks no larger than the configured limit
 
     :param list: a list to break into chunks
     :param limit: maximum size of the returned list
     """
-    for i in range(0, len(list), limit):
-        yield list[i : i + limit]
+    for i in range(0, len(_list), limit):
+        yield _list[i : i + limit]
 
 
 def launch_game_process(instruction: list) -> None:
